Update README.md
Pneumonia poses a serious threat to world health, with large
morbidity and death rates, especially in susceptible groups including children, the elderly, and those with weakened immune
systems. In order to effectively manage pneumonia, a timely
and correct diagnosis is essential. This enables the rapid implementation of suitable treatment options.
A crucial component in confirming the presence of pneumonia has been the use of medical imaging, especially chest
X-rays. Radiologists have historically been the experts in visual interpretation of these pictures, but the complexity of image processing and the growing need for healthcare services
have prompted the investigation of automated techniques. Deep
learning algorithms known as Convolutional Neural Networks
(CNNs) have shown impressive performance in image identification tests, suggesting that they might be a viable option for
improving the effectiveness and precision of pneumonia diagnosis in chest X-rays.